// features/rabbitmq/types.ts
export enum ExchangeType {
  DIRECT = "direct",
  FANOUT = "fanout",
  TOPIC = "topic",
  HEADERS = "headers",
}
// ...
export type QueueNodeData = {
  id: string;
  name: string;
  type: string;
  stats: {
    publish: number;
    ack: number;
  };
  consumer: {
    count: number;
    capacity: number;
  };
};

export type ConsumerNodeData = {
  id: string;
  user: string;
  consumerTag: string;
  count: number;
  activeCount: number;
};

export type ExchangeNodeData = {
  id: string;
  name: string;
  type: ExchangeType;
};

export type BindingNodeData = {
  id: string;
  source: string;
  destination: string;
  routingKeys: string[];
};